pyproject.toml
[project]
name = "cityscape-seg"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Andrew Mitchell", email = "mitchellacoustics15@gmail.com" }
]
dependencies = [
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "transformers>=4.43.2",
    "opencv-python>=4.10.0.84",
    "numpy>=2.0.1",
    "pillow>=10.4.0",
    "h5py>=3.11.0",
    "pandas[performance]>=2.2.2",
    "tqdm>=4.66.4",
    "pyyaml>=6.0.1",
    "scipy>=1.14.0",
]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "BSD-3-Clause" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "mkdocs-material>=9.5.30",
    "mkdocs>=1.6.0",
    "mkdocstrings[python]>=0.25.1",
    "mkdocs-jupyter>=0.24.8",
    "nbconvert>=7.16.4",
    "jupyter>=1.0.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/cityscape_seg"]

[project.scripts]
cityscape-seg = "scripts.run_segmentation:main"
